requirements.txt
pandas
numpy
matplotlib
seaborn
requests
plotly
dash
python-dotenv
scikit-learn
pytest
